Fixed bug by implementing boolean flag named $killPage
Previous update to update username worked correctly when changing both username and email or just the username. But if only the email was changed, the rest of the page showed up after the confirmation page. I fixed this by introducing a boolean variable named $killPage that stops the rest of the HTML code if set to true (it's false by default).