Added employee name to notification email
There is a variable named $fmla_info that is used to concatenate information about the FMLA Request to send in an email. I updated this to include the employee's first and last name with the FMLA request, since the email only addressed the employee by first name, which confused HR. They were not sure which employee submitted this request.